Do not round noisy threshold

+++ src/aggregation/low_count.c
@@ -10,7 +10,7 @@
 typedef struct AidResult
 {
   seed_t aid_seed;
-  int threshold;
+  double threshold;
   bool low_count;
 } AidResult;
